rename source/repos to unask larger questions and conform with gcr package name.
