+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/csr/CloneCloudRepositoryDialog.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.google.cloud.tools.intellij.source.repos.CloneCloudRepositoryDialog">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.google.cloud.tools.intellij.csr.CloneCloudRepositoryDialog">
   <grid id="27dc6" binding="rootPanel" layout-manager="GridLayoutManager" row-count="5" column-count="4" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
@@ -66,7 +66,7 @@
           <text resource-bundle="messages/CloudToolsBundle" key="cloud.repository.dialog.repository.label"/>
         </properties>
       </component>
-      <component id="eae97" class="com.google.cloud.tools.intellij.source.repos.RepositorySelector" binding="repositorySelector" custom-create="true">
+      <component id="eae97" class="com.google.cloud.tools.intellij.csr.RepositorySelector" binding="repositorySelector" custom-create="true">
         <constraints>
           <grid row="1" column="1" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
