+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/csr/RepositoryModelItem.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.source.repos;
+package com.google.cloud.tools.intellij.csr;
 
 import javax.swing.tree.DefaultMutableTreeNode;
