+++ google-cloud-tools-plugin/testSrc/com/google/cloud/tools/intellij/csr/GcpHttpAuthProviderTest.java.excluded
